requirements.txt
fastapi
uvicorn
python-dotenv
bcrypt
openai
stripe
requests
streamlit
